email and username modules working
