+++ frontend/html/Dashboard.html
@@ -46,17 +46,11 @@
 				</div>
 				<div class="modal-body">
 					<!-- USERNAME INPUT -->
-					<div class="mb-3 usernameInput">
+					<div class="mb-3 usernameInputModule">
 
 					</div>
 					<!-- EMAIL INPUT -->
-					<div class="mb-3">
-						<label for="email" class="form-label">Email</label>
-						<input type="email" name="email" class="form-control emailDynamic formInputs"
-							id="emailDashboard" placeholder="What about you ?" autocomplete="off" />
-						<div class="invalid-feedback">
-							Email is not Available!
-						</div>
+					<div class="mb-3 emailInputModule">
 					</div>
 					<!-- BIO INPUT -->
 					<div class="mb-3">
