+++ frontend/js/modules/emailInput.html
@@ -1,0 +1,8 @@
+<label for="email" class="form-label">Email</label>
+<input type="email" name="email" class="emailInput form-control emailDynamic formInputs" placeholder="What about you ?"
+	autocomplete="off" />
+<div class="invalid-feedback">
+	Email is not Available!
+</div>
+
+<script type="module" src="../../js/modules/emailInput.js"></script>
