+++ frontend/js/modules/usernameInput.html
@@ -1,7 +1,7 @@
 <!-- USERNAME INPUT -->
 <div class="mb-3">
 	<label for="usernameDashboard" class="form-label">Username</label>
-	<input type="text" name="username" class="form-control usernameDynamic formInputs" id="usernameDashboard"
+	<input type="text" name="username" class="usernameInput form-control usernameDynamic formInputs"
 		placeholder="Tazz1998" autocomplete="off" />
 	<div class="invalid-feedback">
 		Username is not Available!
@@ -9,4 +9,4 @@
 
 </div>
 
-<script type="module" src="js/modules/usernameInput.js"></script>
+<script type="module" src="../../js/modules/usernameInput.js"></script>
