+++ frontend/js/modules/usernameInput.js
@@ -2,10 +2,10 @@ import { getUserData } from "../User.js";
 import { checkInputAvailable } from "../ApiCalls.js";
 
 var userData = await getUserData();
-var formInputs = document.querySelectorAll(".formInputs");
-formInputs.forEach((input) => {
-	input.addEventListener("input", () => checkUsername(input, userData))
-});
+var module = document.querySelector(".usernameInputModule");
+var input = module.querySelector(".usernameInput");
+
+input.addEventListener("input", () => checkUsername(input, userData));
 
 async function checkUsername(input, userData) {
 	if (input.value !== userData[input.name]) {
@@ -16,8 +16,7 @@ async function checkUsername(input, userData) {
 let debounceTimer;
 
 async function invalidateUsernameIfUnavailable(input, userInput) {
-	var saveChangesButton = document.querySelector(".saveChangesButton");
-	const inputFeedback = document.querySelector('#usernameDashboard ~ div');
+	const inputFeedback = module.querySelector('.usernameInput ~ div');
 	clearTimeout(debounceTimer);
 
 	debounceTimer = setTimeout(async () => {
@@ -25,7 +24,6 @@ async function invalidateUsernameIfUnavailable(input, userInput) {
 			inputFeedback.innerHTML = "Username must be at least 3 characters long";
 			input.classList.remove("is-valid");
 			input.classList.add("is-invalid");
-			saveChangesButton.classList.add("disabled");
 		} else {
 			var usernameAvailable = await checkInputAvailable(input.value, "username");
 			if (input.value == userInput)
@@ -34,12 +32,10 @@ async function invalidateUsernameIfUnavailable(input, userInput) {
 				inputFeedback.innerHTML = "Username is not Available!";
 				input.classList.remove("is-valid");
 				input.classList.add("is-invalid");
-				saveChangesButton.classList.add("disabled");
 
 			} else {
 				input.classList.remove("is-invalid");
 				input.classList.add("is-valid");
-				saveChangesButton.classList.remove("disabled");
 			}
 		}
 	}, 300);
